276 Moves new-subscription-specific component to the PageTemplate folder
